+++ appcontainer/AppContainerModel.js
@@ -4,9 +4,8 @@
  *
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
-import {HoistModel, managed, RootRefreshContextModel, XH} from '@xh/hoist/core';
+import {HoistModel, managed, RootRefreshContextModel, XH, PromiseTaskObserver} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
-import {PendingTaskModel} from '@xh/hoist/utils/async';
 import {AboutDialogModel} from './AboutDialogModel';
 import {BannerSourceModel} from './BannerSourceModel';
 import {ChangelogDialogModel} from './ChangelogDialogModel';
@@ -27,7 +26,7 @@ export class AppContainerModel extends HoistModel {
     // Sub-models
     //------------
     /** Link any async operations that should mask the entire application to this model. */
-    @managed appLoadModel = new PendingTaskModel({mode: 'all'});
+    @managed appLoadModel = new PromiseTaskObserver({mode: 'all'});
 
     @managed aboutDialogModel = new AboutDialogModel();
     @managed changelogDialogModel = new ChangelogDialogModel();