+++ utils/async/index.js
@@ -7,6 +7,3 @@
 
 export * from './AsyncUtils';
 export * from './Timer';
-export * from './task/Task';
-export * from './task/AsyncTask';
-export * from './task/CompoundTask';